requirements.txt
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
numpy==1.23.4
pandas==1.5.0
pdfminer.six==20220524
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
typing-extensions==4.4.0
